Pariam.Investigo.Compact

Description

This module provides utility functions for dealing with clients which request a compact representation of the peer list.

Documentation

ipAndPortToByteString :: String -> Integer -> Maybe ByteString