Pariam.Investigo.Metadata.Types

Description

This module describes types related to how we obtain, store, and manage metadata for a given torrent.

We only care about client data here. "Actual" metadata, such as what a frontend would care about, is stored in another database, but can safely key on our infohash.

The idea is to keep things pluggable, i.e., have the data be able to be stored in any number of ways. As such, we use a free monad and allow users to supply their own interpreter describing how to obtain and update metadata.

Such an interpreter will most likely be a natural transformation of type MetadataOp a -> IO a, but in theory a non-IO implementation could be supplied for testing and so on.

Also, we do not rely on anything in Types, so that we can re-export this module from there.

Documentation

newtype InfoHash

A raw (not URL-encoded) lazy ByteString containing a torrent's infohash.

Constructors

InfoHash ByteString

data Peer

This represents a peer in the verbose (dictionary) representation.

Constructors

Peer

Fields

peerPeerId :: String

peerIp :: String

peerPort :: Integer

data PeerList

The list of peers can be represented compactly or verbosely.

We support both.

Constructors

VerbosePeerList [Peer]

CompactPeerList String

data Metadata

Constructors

Metadata

Fields

metadataInfohash :: String

metadataPeers :: [Peer]

data MetadataOpF md next where

Constructors

MetadataFetch :: InfoHash -> (Maybe md -> next) -> MetadataOpF md next

MetadataUpdate :: (md -> md) -> next -> MetadataOpF md next

metadataUpdate :: forall m md. MonadFree (MetadataOpF md) m => (md -> md) -> m ()

metadataFetch :: forall m md. MonadFree (MetadataOpF md) m => InfoHash -> m (Maybe md)

type MetadataOp = Free (MetadataOpF Metadata)