Pariam.Investigo.Types

Description

This module provides types used throughout Investigo.

Documentation

data BEncodeResponse

This type denotes a response that has been bencoded.

We wrap BEncode so we can create Servant instances on it without the use of orphans, such as MimeRender.

Constructors

BEncodeResponse BEncode

data OneOrZero

Servant will not allow 1 and 0 for Bool fields, it seems. So we write our own type, isomorphic to Bool.

Constructors

One

Zero

data Event

When a client hits the announce url to stop, start, or complete, a torrent, it will include an event queryparam. If it doesn't, or if it's empty, then it is a normal interval-checkin. Otherwise it must be either "started", "stopped", or "completed".

Constructors

Started

Stopped

Completed

data AnnounceRequest

Structure of an announcement request from a client.

Constructors

AnnounceRequest

Fields

infoHash :: Maybe String

peerId :: Maybe String

port :: Maybe Integer

uploaded :: Maybe Integer

downloaded :: Maybe Integer

left :: Maybe Integer

compact :: Maybe OneOrZero

noPeerId :: Maybe OneOrZero

event :: Maybe Event

ip :: Maybe String

numwant :: Maybe Integer

key :: Maybe String

trackerid :: Maybe String

data AnnounceResponse

Response to announce request.

Constructors

Failure

Fields

failureReason :: String

type API = "announce" :> (QueryParam "info_hash" String :> (QueryParam "peer_id" String :> (QueryParam "port" Integer :> (QueryParam "uploaded" Integer :> (QueryParam "downloaded" Integer :> (QueryParam "left" Integer :> (QueryParam "compact" OneOrZero :> (QueryParam "no_peer_id" OneOrZero :> (QueryParam "event" Event :> (QueryParam "ip" String :> (QueryParam "numwant" Integer :> (QueryParam "key" String :> (QueryParam "trackerid" String :> Get '[PlainText] BEncodeResponse)))))))))))))